requirements.txt
transformers>=4.25.0
pillow>=9.0.0
diffusers>=0.12.0
torch>=1.12.0